rice.awd.shb: Reproducible Research Compendium for Analysing Effects of Water Management and Nitrogen on Rice Sheath Blight

Provides a research compendium of analysis associated with experiments conducted at the International Rice Research Institute (IRRI), Los Baños, Calabarzon, Philippines in the dry seasons of 2015 and 2016. The experiments were designed to test the effects of alternate wetting and drying irrigation management on rice sheath blight disease. The plots were split plot with an irrigation main plot and nitrogen sub-plot treatments.

See also

Author

Maintainer: Adam H. Sparks adamhsparks@gmail.com (ORCID)

Authors:

Nancy P. Castilla n.castilla@irri.org

Bjoern Ole Sander b.sander@irri.org (ORCID)

Michael Noel

Other contributors:

Global Rice Science Partnership (GRiSP) [funder]